_original/pyproject.toml
[tool.poetry]
name = "scssg"
version = "0.1.0"
description = ""
authors = ["Noah Roberts <noah@noahsroberts.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
markdown2 = "^2.5.3"
jinja2 = "^3.1.5"
pyyaml = "^6.0.2"
python-frontmatter = "^1.1.0"

[tool.poetry.scripts]
cli = "scssg.__main__:cli"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
